pyproject.toml
[project]
name = "cc-compse"
version = "0.1.0"
description = "Cloud Computing CS{16,20}60 docker compose project"
authors = [
    {name = "dansc0de dpm79@pitt.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2-binary (>=2.9.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
